requirements.txt
Pillow
mysql
numpy
pathlib 
open3d
matplotlib
opencv-python
scikit-image
imutils
scipy
Langchain
streamlit
openai
-e .